+++ Catala/PHP/Directori/DirectoriUpdate.php
@@ -21,9 +21,10 @@ $condIMG = "";
 if ($IMG)
 {
 	$SQL = "SELECT Imatge FROM Directori WHERE IdDirectori = ".$id;	
-	$result = mysql_query($SQL,$oConn);
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
 
-	while ($row = mysql_fetch_array($result))
+
+	 while ($row = $result->fetch_assoc())
 	{
 		if ($row["Imatge"]){
 			if(file_exists("../../../imgDirectori/".$row["Imatge"]))unlink("../../../imgDirectori/".$row["Imatge"]);
